회원 항목이랑 맞춰 Long -> int로 전부 교체 / 로그인 후 글작성 확인 / 로그인 된 멤버로 삭제 확인 / 로그인 후 노출되야하는 버튼들 확인 (사진은 어떤때는 나오고 어떤떄는 안나오는데 서버설정문제 같음..)

+++ CarTalkProject/src/main/java/com/kh/spring/used/model/dto/CarInfoDTO.java
@@ -16,8 +16,8 @@ import lombok.ToString;
 @ToString
 public class CarInfoDTO {
 	
-	private Long carNo;
-	private Long usedNo;
+	private int carNo;
+	private int usedNo;
 	private String manufacturer;
 	private String model;
 	private String carYear;
